requirements.txt
openenv-core>=0.2.2
groq>=0.4.2
gradio>=6.0.0
pandas>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
scikit-learn>=1.0.0
